refactor: "나누기 소숫점 구현"
나누기 할때 소숫점 자릿수 구현 및 일반 나머지 표현, 줄바꿈 및 엔터 등 편의성 개선

#

+++ Lv2Calculator/DivideOperation.java
@@ -3,6 +3,6 @@ package Lv2Calculator;
 public class DivideOperation extends AbsOperation {
     @Override
     public double operate(int firstNumber, int secondNumber) {
-        return firstNumber / secondNumber;
+        return (double) firstNumber / secondNumber;
     }
 }
